Add files via upload
Cyber attacks are in the news on a daily basis, impacting individuals and organizations, whether in the public or private sector. In this lesson, we will review three high profile case studies of cyber attacks so you can understand the extent of what is possible and going on in the world. Each case study focuses on a different type of threat actor. These three case studies are part of an ever-growing catalogue of security breaches in the international landscape. As a participant within the security community, it is important to learn from examples to guide future decision making.